push

+++ 0x17-doubly_linked_lists/7-insert_dnodeint.c
@@ -56,7 +56,7 @@ dlistint_t *insert_dnodeint_at_index(dlistint_t **h, unsigned int idx, int n)
 {
 	dlistint_t *new_node = NULL;
 	dlistint_t *position = NULL;
-	size_t size = dlistint_size(*h);
+	unsigned int size = (unsigned int)dlistint_size(*h);
 
 	if (!h || idx > size || idx < 0)
 		return (NULL);
